// src/types.ts
export class Annotation {
    //todo: check types
    indexCount: number;
    bookName: string;
    bookPath: string;
    lowerCasePath: string;
    sectionNumber: number;
    placeholder1: string;
    location: string;
    characterCount: number;
    signedColor: number;
    unixTimestamp: string;
    bookmarkText: string;
    noteText: string;
    highlightText: string;
    annotType1: number;
    annotType2: number;
    annotType3: number;
}